+++ contracts/wallet/src/schema.rs
@@ -63,54 +63,3 @@ impl WalletOffchainInput {
         serde_json::to_string(self).expect("JSON: failed to serialize")
     }
 }
-
-// // TODO: docs
-// #[cfg_attr(
-//     feature = "serde",
-//     derive(::serde::Serialize, ::serde::Deserialize),
-//     cfg_attr(feature = "schemars-v0_8", derive(::schemars::JsonSchema))
-// )]
-// #[cfg_attr(feature = "arbitrary", derive(::arbitrary::Arbitrary))]
-// #[derive(Debug, Clone, PartialEq, Eq, Hash)]
-// pub struct WalletOffchainMessage {
-//     pub output: String,
-
-//     // TODO: maybe only if self?
-//     pub msg: OffchainMessage,
-
-//     #[cfg_attr(
-//         feature = "serde",
-//         serde(default, skip_serializing_if = "Option::is_none")
-//     )]
-//     pub as_extension_id: Option<AccountId>,
-
-//     pub proof: Proof,
-// }
-
-// impl WalletOffchainMessage {
-//     #[inline]
-//     pub fn as_self(proof: impl Into<Proof>) -> Self {
-//         Self {
-//             as_extension_id: None,
-//             proof: proof.into(),
-//         }
-//     }
-
-//     #[inline]
-//     pub fn as_extension(extension_id: impl Into<AccountId>, proof: impl Into<Proof>) -> Self {
-//         Self {
-//             as_extension_id: Some(extension_id.into()),
-//             proof: proof.into(),
-//         }
-//     }
-
-//     // TODO: naming
-//     #[cfg(feature = "json")]
-//     #[inline]
-//     pub fn wrap_as_extension(self, extension_id: impl Into<AccountId>) -> Self {
-//         Self::as_extension(
-//             extension_id,
-//             serde_json::to_string(&self).expect("JSON: failed to serialize"),
-//         )
-//     }
-// }
